:recycle: 체크박스 컴포넌트 리팩토링
